splits wsgi and wiznet eth
