+++ libs/w5XXXX-evb-pico/wiznet5keth.py
@@ -5,7 +5,6 @@ Requires the following circuit python libs:
 
     - adafruit_wiznet5k
     - adafruit_bus_device
-    - adafruit_requests.mpy
 
 """
 
@@ -16,13 +15,6 @@ import time
 
 # Wiznet ethernet WIZNET5K, etc
 from adafruit_wiznet5k.adafruit_wiznet5k import *
-
-# Wiznet WSGI server
-import adafruit_requests as requests
-from adafruit_wsgi.wsgi_app import WSGIApp
-import adafruit_wiznet5k.adafruit_wiznet5k_wsgiserver as server
-import adafruit_wiznet5k.adafruit_wiznet5k_socket as socket
-
 
 ##SPI0
 SPI0_SCK = board.GP18
@@ -87,16 +79,3 @@ def config_eth(netconfig, debug=False):
     print("IP address:", eth.pretty_ip(eth.ip_address))
     # Return initialized WIZNET5K ethernet
     return eth
-
-
-def wsgi_web_server(eth, listening_port=80):
-    """Returns a a configured WSGI server and web app object.
-
-    Params:
-      - eth (WIZNET5K): A configured ethernet object. Returned from config_eth()
-      - listening_port (int): The port for the web app to listen on
-    """
-    requests.set_socket(socket, eth)  # Initialize a requests object with a socket and ethernet interface
-    server.set_interface(eth)
-    web_app = WSGIApp()  # WSGI web app
-    return (server.WSGIServer(listening_port, application=web_app), web_app)
